Add core tactics
The core axioms and rules were found by searching for every occurrence
of coreAxiom(...) and coreRule(...) in Ax.scala.

The implementations of RenUSubst.toHippo are mostly redundant and
probably boil down to two cases if you look at them angrily enough.
Maybe worth a future refactoring, but it's working fine for now.