Remove DerivedHippoProof
A DerivedHippoProof represents a uniquely identifiable proof rule
(conclusion and premises) plus a way to compute a proof for said rule
from a HippoContext. Since the rule is uniquely identifiable, it can be
used to cache the proof.

Since hippolib can now use `lazy val` to defer proof computation, there
is no longer a need to separate a rule and its proof. However, there is
still a need to cache the proofs so they are not recomputed on every
restart. For this, a more general tactic-based caching system will be
used.

In conclusion, all of DerivedHippoProof's functions have been or will be
fulfilled by different systems, hence DerivedHippoProof itself is no
longer necessary.