Hash TacticInfo
Since a TacticConstructor is effectively a Scala function, there's no
obvious way to hash it even though it *should* normally be pure.

In the case of tactics, the tactic class name (and its classloader's
name) are usually hashed along with the arguments, but since
TacticConstructors are often anonymous classes based on lambdas, this
isn't really an option.

Instead, the user needs to manually provide a globally unique name for
the TacticConstructor that is then hashed. A globally unique id would
also suffice, but it would be a bit more annoying to assign when
creating a new tactic and might lead to merge conflicts more often.