Improve error for import in pure context
In order to show which import is not allowed, HippoExpression.Import
needs to know where it was defined. Of course this may not be possible
in some scenarios. Maybe in those cases, a dummy SourceFile and Slice
can be supplied? At this point, I'd rather not make all slices optional
because it would complicate a few things without immediate benefit.