Add BelleValue and use it instead of HippoValue in hippolib
Hippolib should be independent of hippolang, and HippoValue did not have
the ability to represent all kinds of values one might want to pass to a
bellerophon tactic.

Using Any instead would not work because Any can't be hashed, and these
values need to be hashable because they're the arguments to tactics.